update du lieu

+++ MTshop/database/migrations/2019_11_29_070618_create_cusc_mau_table.php
@@ -15,7 +15,7 @@ class CreateCuscMauTable extends Migration
     {
         Schema::create('cusc_mau', function (Blueprint $table) {
             $table->engine = 'InnoDB';
-            $table->unsignedTinyInteger('m_ma')->autoIncrement()->comment('Mã màu sản phẩm, 1-Phối màu (đỏ, vàng, ...)');
+            $table->unsignedBigInteger('m_ma')->autoIncrement()->comment('Mã màu sản phẩm, 1-Phối màu (đỏ, vàng, ...)');
             $table->string('m_ten', 50)->comment('Tên màu # Tên màu sản phẩm');
             $table->timestamp('m_taoMoi')->default(DB::raw('CURRENT_TIMESTAMP'))->comment('Thời điểm tạo # Thời điểm đầu tiên tạo màu');
             $table->timestamp('m_capNhat')->default(DB::raw('CURRENT_TIMESTAMP'))->comment('Thời điểm cập nhật # Thời điểm cập nhật màu gần nhất');
